[Engineering] Widen runs.rows_loaded to BigInt (closes #283)
Third of the V2 P5 int32-overflow fixes — same shape as b8c3d4e5f6g7
(usage_ledger.quantity, core#272) and c9d4e5f6g7h8 (notifications.type
width, core#279). Surfaced by QA's core#282 realistic-size probe
which had been landed as an xfail against 3 billion rows pointing at
this issue.

Enterprise customers routinely run one-shot backfills of clickstream,
event, and log data that exceed 2^31 rows in a single pipeline run.
``dlt_runner._extract_rows_loaded`` returns the dlt row count,
``upload_tasks`` passes it to ``execution_service.complete_run``, and
``complete_run`` writes ``run.rows_loaded = rows_loaded``. On any
count ≥ 2^31, Postgres raises ``NumericValueOutOfRange: integer out of
range`` on the INSERT and the run record goes unpersisted.

Fix:
- Migration ``d0e5f6g7h8i9`` runs ``ALTER TABLE runs ALTER COLUMN
  rows_loaded TYPE bigint``. Metadata-only on Postgres — no table
  rewrite. Safe to run live.
- Model ``Run.rows_loaded`` now declares ``BigInteger``. Dropped the
  unused ``Integer`` import.

Audit of other ``runs`` columns (documented in migration docstring):
- ``id``: Integer autoincrement PK — ~10yrs headroom at current
  trajectory, widening deferred.
- ``target_id`` / ``org_id``: already BigInteger ✓
- ``uploaded_files.file_size``: already BigInteger ✓ (u0q7r8s9t1n2)

Tests:
- 4 new regressions in ``test_d0e5_bigint_rows_loaded.py``: revision
  id + upgrade/downgrade shape + model column type assert.
- Flipped QA's ``test_runs_rows_loaded_bigint_roundtrip`` xfail in
  ``test_roundtrip.py`` — it ships as a passing 3-billion-row probe
  now, catching any future reversion back to int32.

2142 core tests pass. Ruff + i18n parity clean.

+++ datanika/models/run.py
@@ -1,7 +1,7 @@
 import enum
 from datetime import datetime
 
-from sqlalchemy import BigInteger, DateTime, Enum, Integer, Text
+from sqlalchemy import BigInteger, DateTime, Enum, Text
 from sqlalchemy.orm import Mapped, mapped_column
 
 from datanika.models.base import Base, TenantMixin, TimestampMixin
@@ -32,5 +32,9 @@ class Run(Base, TenantMixin, TimestampMixin):
     started_at: Mapped[datetime | None] = mapped_column(DateTime(timezone=True), nullable=True)
     finished_at: Mapped[datetime | None] = mapped_column(DateTime(timezone=True), nullable=True)
     logs: Mapped[str | None] = mapped_column(Text, nullable=True)
-    rows_loaded: Mapped[int | None] = mapped_column(Integer, nullable=True)
+    # BigInteger: Enterprise clickstream/event/log backfills routinely
+    # exceed 2^31 rows in a single pipeline run. int32 would overflow on
+    # insert with ``NumericValueOutOfRange``. See core#283 — same class
+    # as the usage_ledger.quantity widening in core#272.
+    rows_loaded: Mapped[int | None] = mapped_column(BigInteger, nullable=True)
     error_message: Mapped[str | None] = mapped_column(Text, nullable=True)
